Move helper methods to dialog overlay
